+++ src/aks-preview/azext_aks_preview/_validators.py
@@ -32,6 +32,8 @@ from azext_aks_preview._consts import (
     CONST_OUTBOUND_TYPE_USER_DEFINED_ROUTING,
     CONST_AZURE_SERVICE_MESH_INGRESS_MODE_EXTERNAL,
     CONST_AZURE_SERVICE_MESH_INGRESS_MODE_INTERNAL,
+    CONST_MANAGED_CLUSTER_SKU_TIER_FREE,
+    CONST_MANAGED_CLUSTER_SKU_TIER_STANDARD,
 )
 from azext_aks_preview._helpers import _fuzzy_match
 
@@ -210,6 +212,15 @@ def validate_load_balancer_outbound_ips(namespace):
         if not all(ip_id_list):
             raise CLIError(
                 "--load-balancer-outbound-ips cannot contain whitespace")
+
+
+def validate_sku_tier(namespace):
+    """Validates the sku tier string."""
+    if namespace.tier is not None:
+        if namespace.tier == '':
+            return
+        if namespace.tier.lower() not in (CONST_MANAGED_CLUSTER_SKU_TIER_FREE, CONST_MANAGED_CLUSTER_SKU_TIER_STANDARD):
+            raise InvalidArgumentValueError("--tier can only be free or standard")
 
 
 def validate_load_balancer_outbound_ip_prefixes(namespace):